LayoutTests/imported/w3c/IndexedDB-private-browsing/idbobjectstore_get-expected.txt


PASS IDBObjectStore.get() - key is a number
